Moved script modules to Modules folder except driver script
